Fix NaN detection by the GPU on float textures for window platform
float textures coming from BITPIX<0 fits images are sent to GPU as
RGBA8UI textures. Float decoding from a vec4 rgba is done in the shader.
Then this obtained decoded float can be tested against nan/inf in the
shader

+++ src/glsl/webgl2/fits/base.vert
@@ -1,6 +1,6 @@
 #version 300 es
-precision lowp float;
-precision mediump int;
+precision highp float;
+precision highp int;
 
 layout (location = 0) in vec2 ndc_pos;
 layout (location = 1) in vec2 uv;
